pyproject.toml
[project]
name = "vm-transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20240930",
    "streamlit>=1.44.1",
]
